gmail-subscription-tracker/requirements.txt
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
beautifulsoup4==4.12.3
tabulate==0.9.0
